refactor: move useComment
